Changed engine.py
Removed second trailing underscore from `_SCL`, `_RL` and `_MSG` variables as this led to interesting errors while testing.
Reformatted file to match python standards.

Reworked the `message_load_file` function as it led to bugs and could not handle headers included into the file. The new implementation can handle this and has no problems with special, non-ascii characters anymore.
Added some exception handling to the `smtp.login` call in `send_message` function. There is now also a TODO for handling possibly occurring ssl errors when the user tries to use encrypted connections on servers that do not support this.
Added 'To' headers in `send_message`, so the recipient does not only get the message but is also shown as the recipient in the mail client.